+++ frontend/src/pages/Profile/ProfileHeader.tsx
@@ -1,8 +1,10 @@
 import React from "react";
-import { Grid, Typography } from "@mui/material";
+import { Button, Grid, Typography } from "@mui/material";
 // @ts-ignore
 import styles from "./Profile.module.css";
 import { ProfileDto } from "../../types/profile/profile.dto";
+import { useDispatch } from "react-redux";
+import { fetchedStatus } from "../../redux/ducks/profile/profile.slice";
 
 
 interface ProfileInfo {
@@ -10,6 +12,11 @@ interface ProfileInfo {
 }
 
 export const ProfileHeader: React.FC<ProfileInfo> = ({ profile }) => {
+
+	const dispatch = useDispatch();
+
+	const onStatusChange = () => dispatch(fetchedStatus( { id: profile?.id, status: 'The capitalist himself will sell the rope on which he will be hanged'}))
+
 	return (
 		<Grid>
 			{!profile?.activated &&
@@ -20,7 +27,12 @@ export const ProfileHeader: React.FC<ProfileInfo> = ({ profile }) => {
 				<img src={profile?.avatar} className={styles.profileAvatar} />
 				<Grid>
 					<Typography variant="h5">{profile?.fullname}</Typography>
-					<Typography variant="h6">{profile?.username}</Typography>
+					<Grid className={styles.flex}>
+						<Grid className={styles.status}>
+							<Typography variant="subtitle2">{profile?.status}</Typography>
+						</Grid>
+						<Button className={styles.button} variant="contained" onClick={onStatusChange}>Change status</Button>
+					</Grid>
 				</Grid>
 			</Grid>
 		</Grid>